framework/mlops/requirements-mlops.txt
databricks-sql-connector
pandas
scikit-learn
joblib